Some Change FOr Private Routing

+++ src/Controller/Cada/CadaDeleteById.js
@@ -3,12 +3,16 @@ const { Cada } = require("../../Dbconfig/DatabaseConfig");
 
 const cadaDeleteById = async (req, res, next) => {
   try {
+    const email = req?.decoded?.email;
+    if (!email) {
+      return;
+    }
     const id = req.params.id;
     const deleteValue = await Cada.findOneAndDelete({ _id: new ObjectId(id) });
     return handleSuccess(res, {
       statusCode: 200,
       message: "সফলভাবে তথ্যটি ডিলেট করা হয়েছে",
-      payload:  deleteValue ,
+      payload: deleteValue,
     });
   } catch (error) {
     next(error);
